+++ src/main/java/com/natoboram/switcheroo/Main.java
@@ -25,6 +25,8 @@ public class Main implements ClientModInitializer {
 	 * This logger is used to write text to the console and the log file. It is
 	 * considered best practice to use your mod id as the logger's name. That way,
 	 * it's clear which mod wrote info, warnings, and errors.
+	 * <p>
+	 * Note: {@link Logger#debug} does not write to the console.
 	 */
 	private final static Logger LOGGER = LogManager.getLogger(MOD_ID);