+++ src/main/java/com/natoboram/switcheroo/Switch.java
@@ -1,5 +1,8 @@
 package com.natoboram.switcheroo;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.MinecraftClient;
@@ -10,6 +13,7 @@ import net.minecraft.item.ItemStack;
 @Environment(EnvType.CLIENT)
 public class Switch {
 
+	private static final Logger LOGGER = LogManager.getLogger(Main.MOD_ID);
 	private static final MinecraftClient CLIENT = MinecraftClient.getInstance();
 
 	/**
@@ -18,16 +22,15 @@ public class Switch {
 	 * @param player PlayerEntity that's about to execute the switcheroo.
 	 * @param item   Item that should be put in its hand.
 	 */
-	public static void switcheroo(final PlayerEntity player, final ItemStack item) {
+	public static void switcheroo(final PlayerEntity player, final ItemStack item, final SwitcherooConfig config) {
 		final PlayerInventory inventory = player.getInventory();
-
-		// Don't send useless packets
-		if (inventory.getMainHandStack().isItemEqualIgnoreDamage(item))
-			return;
 
 		final int slot = inventory.getSlotWithStack(item);
 		if (slot == -1)
 			return;
+
+		if (config.debug)
+			LOGGER.info("Switching for " + item);
 
 		if (PlayerInventory.isValidHotbarIndex(slot)) {
 			// Select the item from the hotbar